requirements.txt
streamlit==1.15.0
tensorflow==2.11.0
numpy==1.21.2
Pillow==8.2.0
tensorflow_hub==0.12.0  # Only if you are using TensorFlow Hub layers
